DOCS: revised the structure of the html directory to avoid redirection pages and to keep previous versions

+++ tests/conftest.py
@@ -5,6 +5,7 @@
 # ======================================================================================
 # ruff: noqa
 
+import os
 import pathlib
 
 import numpy as np
@@ -35,13 +36,16 @@ def pytest_sessionfinish(session, exitstatus):  # pragma: no cover
     for f in list(cwd.glob("**/*.jdx")):
         f.unlink()
     for f in list(cwd.glob("**/*.json")):
-        if f.name != "zenodo.json":
+        if (
+            f.name
+            not in [
+                "zenodo.json",
+            ]
+            and ".vscode" not in f.parts
+            and ".venv" not in f.parts
+        ):
             f.unlink()
     for f in list(cwd.glob("**/*.log")):
-        f.unlink()
-
-    docs = cwd / "docs"
-    for f in list(docs.glob("**/*.ipynb")):
         f.unlink()
